simplify init parameter of FaDIn and move Hawkes initialization to a separate file
