+++ fadin/utils/utils.py
@@ -1,6 +1,5 @@
 import numpy as np
 import torch
-import matplotlib.pyplot as plt
 
 
 def kernel_normalization(kernel_values, time_values, delta, lower=0, upper=1):
@@ -128,114 +127,3 @@ def check_random_state(seed):
         return seed
     raise ValueError('%r cannot be used to seed a numpy.random.RandomState'
                      ' instance' % seed)
-
-
-def momentmatching_kernel_nomark(solver, events, n_ground_events,
-                                 plot_delta=False, mode='max'):
-    """Moment matching initialization of kernel parameters. Implemented for
-    'truncated_gaussian' and 'raised_cosine' kernels.
-    For the truncated gaussian kernel, the means $m_{i,j}$ and std
-    $\\sigma_{i,j}$ are:
-    $m_{i, j} =
-    \\frac{1}{N_{g_i}(T)}\\sum_{t_n^i \\in \\mathscr{F}_T^i}
-    \\delta t^{i, j}_n$
-    $\\sigma_{i, j} =
-    \\sqrt{\\dfrac{
-        \\sum_{t_n^i \\in \\mathscr{F}_T^i} (\\delta t^{i, j}_n - m_{i, j})^2
-    }{N_{g_i}(T) - 1}}.
-    For the raised cosine kernel, the parameters $u_{i,j}$ and $s_{i,j} are:
-    $u^{\\text{m}}_{i, j} =
-    \\max{(0, m^{\\text{m}}_{i, j} - \\sigma^{\\text{m}}_{i, j})}$
-    $s^{\\text{m}}_{i, j} = \\sigma_{i, j}^{\\text{m}}$
-
-    Parameters
-    ----------
-    solver :  `FaDIn` or `MarkedFaDIn` object
-        The solver object
-    events : list of torch.tensor
-        List of events for each dimension
-    n_ground_events : torch.tensor
-        Number of ground events for each dimension
-    plot_delta : bool, default=False
-        Whether to plot the delta_t distribution
-    mode : str, default='max'
-        Mode to compute the delta_t distribution. Supported values are 'max'
-        and 'mean'.
-
-    Returns
-    -------
-    list of torch.tensor
-        List of kernel parameters
-
-    """
-    kernel_params_init = [torch.ones(solver.n_dim, solver.n_dim),
-                          torch.ones(solver.n_dim, solver.n_dim)]
-    for i in range(solver.n_dim):
-        for j in range(solver.n_dim):
-            # Mean, std of time delta of [i, j] kernel
-            if mode == 'max':
-                delta_t = torch.zeros(int(n_ground_events[i].item()))
-                for n in range(int(n_ground_events[i])):
-                    t_n_i = events[i][n]
-                    t_n_j = torch.max(
-                        torch.where(torch.tensor(events[j]) < t_n_i,
-                                    torch.tensor(events[j]),
-                                    0.)
-                        )
-                    delta_t[n] = t_n_i - t_n_j
-                avg = torch.mean(delta_t)
-                std = torch.std(delta_t)
-            if mode == 'mean':
-                delta_t = []
-                for n in range(int(n_ground_events[i])):
-                    t_n_i = events[i][n]
-                    for t_n_j in events[j]:
-                        if t_n_j < t_n_i - solver.W:
-                            continue
-                        if t_n_j >= t_n_i:
-                            break
-                        delta_t.append(t_n_i - t_n_j)
-                avg = torch.mean(torch.tensor(delta_t))
-                std = torch.std(torch.tensor(delta_t))
-            # Plot delta_t distribution
-            if plot_delta:
-                fig_delta, ax_delta = plt.subplots()
-                ax_delta.hist(delta_t, bins=20)
-                ax_delta.set_xlim([0, solver.W])
-                ax_delta.set_xlabel('Time')
-                ax_delta.set_ylabel('Histogram')
-                fig_delta.suptitle('Moment Matching delta_t')
-                fig_delta.show()
-            # Parameters of [i, j] kernel
-            if solver.kernel == 'truncated_gaussian':
-                kernel_params_init[0][i, j] = avg
-            if solver.kernel == 'raised_cosine':
-                u = max(avg - std, 0)
-                kernel_params_init[0][i, j] = u
-            kernel_params_init[1][i, j] = std
-    return kernel_params_init
-
-
-def momentmatching_nomark(solver, events, n_ground_events, end_time,
-                          mode='max'):
-    """Moment matching initialization of baseline, alpha and kernel parameters.
-    $\\mu_i^s = \frac{\\#\\mathscr{F}^i_T}{(D+1)T} \forall i \\in [1, D]$
-    $\\alpha_{i, j}^s = \\frac{1}{D+1} \\forall i,j \\in [1, D]$
-    Kernel parameters are initialized by `momentmatching_kernel`.
-    """
-    assert solver.kernel in ['truncated_gaussian', 'raised_cosine'], (
-        f"Smart initialization not implemented for kernel {solver.kernel}"
-    )
-    # Baseline init
-    baseline = n_ground_events / (end_time * (solver.n_dim + 1))
-    baseline = (baseline * solver.baseline_mask).requires_grad_(True)
-
-    # Alpha init
-    alpha = torch.ones(solver.n_dim, solver.n_dim) / (solver.n_dim + 1)
-    alpha = (alpha * solver.alpha_mask).requires_grad_(True)
-
-    # Kernel parameters init
-    kernel_params_init = momentmatching_kernel_nomark(
-        solver, events, n_ground_events, mode=mode
-    )
-    return baseline, alpha, kernel_params_init
